Migrate Series to Indexed in additional components
- Continued the transition by replacing occurrences of `Series<T>` with `Indexed<T>` in various files, enhancing code clarity and performance.
- Adjusted methods and data structures to utilize the `Indexed` alias, ensuring compatibility with existing functionality.
- This change supports the lazy migration approach, facilitating a smooth transition without disrupting existing operations.

These updates contribute to the overall improvement of the codebase structure and further support ongoing migration efforts.

+++ Trikeshed/src/commonMain/kotlin/borg/trikeshed/util/Sha256Common.kt
@@ -4,28 +4,28 @@ package borg.trikeshed.util
 // Adapted from public domain sources, minimal and direct
 
 private val K = intArrayOf(
-    0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1, 0x923f82a4, 0xab1c5ed5,
-    0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3, 0x72be5d74, 0x80deb1fe, 0x9bdc06a7, 0xc19bf174,
-    0xe49b69c1, 0xefbe4786, 0x0fc19dc6, 0x240ca1cc, 0x2de92c6f, 0x4a7484aa, 0x5cb0a9dc, 0x76f988da,
-    0x983e5152, 0xa831c66d, 0xb00327c8, 0xbf597fc7, 0xc6e00bf3, 0xd5a79147, 0x06ca6351, 0x14292967,
-    0x27b70a85, 0x2e1b2138, 0x4d2c6dfc, 0x53380d13, 0x650a7354, 0x766a0abb, 0x81c2c92e, 0x92722c85,
-    0xa2bfe8a1, 0xa81a664b, 0xc24b8b70, 0xc76c51a3, 0xd192e819, 0xd6990624, 0xf40e3585, 0x106aa070,
-    0x19a4c116, 0x1e376c08, 0x2748774c, 0x34b0bcb5, 0x391c0cb3, 0x4ed8aa4a, 0x5b9cca4f, 0x682e6ff3,
-    0x748f82ee, 0x78a5636f, 0x84c87814, 0x8cc70208, 0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2
+    0x428a2f98.toInt(), 0x71374491.toInt(), 0xb5c0fbcf.toInt(), 0xe9b5dba5.toInt(), 0x3956c25b.toInt(), 0x59f111f1.toInt(), 0x923f82a4.toInt(), 0xab1c5ed5.toInt(),
+    0xd807aa98.toInt(), 0x12835b01.toInt(), 0x243185be.toInt(), 0x550c7dc3.toInt(), 0x72be5d74.toInt(), 0x80deb1fe.toInt(), 0x9bdc06a7.toInt(), 0xc19bf174.toInt(),
+    0xe49b69c1.toInt(), 0xefbe4786.toInt(), 0x0fc19dc6.toInt(), 0x240ca1cc.toInt(), 0x2de92c6f.toInt(), 0x4a7484aa.toInt(), 0x5cb0a9dc.toInt(), 0x76f988da.toInt(),
+    0x983e5152.toInt(), 0xa831c66d.toInt(), 0xb00327c8.toInt(), 0xbf597fc7.toInt(), 0xc6e00bf3.toInt(), 0xd5a79147.toInt(), 0x06ca6351.toInt(), 0x14292967.toInt(),
+    0x27b70a85.toInt(), 0x2e1b2138.toInt(), 0x4d2c6dfc.toInt(), 0x53380d13.toInt(), 0x650a7354.toInt(), 0x766a0abb.toInt(), 0x81c2c92e.toInt(), 0x92722c85.toInt(),
+    0xa2bfe8a1.toInt(), 0xa81a664b.toInt(), 0xc24b8b70.toInt(), 0xc76c51a3.toInt(), 0xd192e819.toInt(), 0xd6990624.toInt(), 0xf40e3585.toInt(), 0x106aa070.toInt(),
+    0x19a4c116.toInt(), 0x1e376c08.toInt(), 0x2748774c.toInt(), 0x34b0bcb5.toInt(), 0x391c0cb3.toInt(), 0x4ed8aa4a.toInt(), 0x5b9cca4f.toInt(), 0x682e6ff3.toInt(),
+    0x748f82ee.toInt(), 0x78a5636f.toInt(), 0x84c87814.toInt(), 0x8cc70208.toInt(), 0x90befffa.toInt(), 0xa4506ceb.toInt(), 0xbef9a3f7.toInt(), 0xc67178f2.toInt()
 )
 
 fun sha256(input: String): String {
     val bytes = input.encodeToByteArray()
     val padded = pad(bytes)
     val blocks = padded.size / 64
-    var h0 = 0x6a09e667
-    var h1 = 0xbb67ae85
-    var h2 = 0x3c6ef372
-    var h3 = 0xa54ff53a
-    var h4 = 0x510e527f
-    var h5 = 0x9b05688c
-    var h6 = 0x1f83d9ab
-    var h7 = 0x5be0cd19
+    var h0 = 0x6a09e667.toInt()
+    var h1 = 0xbb67ae85.toInt()
+    var h2 = 0x3c6ef372.toInt()
+    var h3 = 0xa54ff53a.toInt()
+    var h4 = 0x510e527f.toInt()
+    var h5 = 0x9b05688c.toInt()
+    var h6 = 0x1f83d9ab.toInt()
+    var h7 = 0x5be0cd19.toInt()
     val w = IntArray(64)
     for (i in 0 until blocks) {
         val base = i * 64
